Refactor: SC,CC 분리 및 SEO최적화#3
