+++ src/components/board/CCBoardCategory.tsx
@@ -1,0 +1,59 @@
+"use client";
+
+import { useRouter, useSearchParams } from "next/navigation";
+import DefaultButton from "@/components/ui/defaultButton";
+
+const categories = [
+  "전체",
+  "공지사항",
+  "보안분석",
+  "기술자료",
+  "스터디",
+  "자유게시판",
+];
+
+interface BoardCategoryProps {
+  selectedCategory: string;
+}
+
+export default function BoardCategory({
+  selectedCategory,
+}: BoardCategoryProps) {
+  const router = useRouter();
+  const searchParams = useSearchParams();
+
+  const handleCategoryChange = (category: string) => {
+    const params = new URLSearchParams(searchParams);
+
+    if (category === "전체") {
+      params.delete("category");
+    } else {
+      params.set("category", category);
+    }
+    params.delete("page"); // 카테고리 변경 시 첫 페이지로
+
+    const queryString = params.toString();
+    const newUrl = queryString ? `/board?${queryString}` : "/board";
+    router.push(newUrl);
+  };
+
+  return (
+    <div className="flex gap-2 flex-wrap">
+      {categories.map((category) => (
+        <DefaultButton
+          key={category}
+          variant={selectedCategory === category ? "default" : "outline"}
+          size="sm"
+          onClick={() => handleCategoryChange(category)}
+          className={
+            selectedCategory === category
+              ? "bg-cert-red text-white hover:bg-cert-red/80"
+              : "border-gray-300 text-gray-600 hover:border-cert-red"
+          }
+        >
+          {category}
+        </DefaultButton>
+      ))}
+    </div>
+  );
+}
