+++ src/components/board/SCBoardCardList.tsx
@@ -1,0 +1,41 @@
+import BoardCard from "@/components/board/SCBoardCard";
+import DefaultNoneResultUi from "@/components/ui/defaultNoneResultUi";
+import ShieldSVG from "@/icons/shield.svg";
+
+interface BoardContent {
+  id: number;
+  title: string;
+  content: string;
+  author: string;
+  date: string;
+  category: string;
+  views: number;
+  likes: number;
+  comments: number;
+  isNotice: boolean;
+  priority: string;
+}
+
+export default function BoardCardList({
+  contents,
+}: {
+  contents: BoardContent[];
+}) {
+  if (contents.length === 0) {
+    return (
+      <DefaultNoneResultUi
+        icon={<ShieldSVG />}
+        title="검색 결과가 없습니다."
+        description="검색어를 확인하거나 다른 카테고리를 선택해보세요."
+      />
+    );
+  }
+
+  return (
+    <div className="flex flex-col space-y-4">
+      {contents.map((content) => (
+        <BoardCard key={content.id} {...content} />
+      ))}
+    </div>
+  );
+}
