--- src/components/board/boardPagination.tsx
@@ -1,106 +1,0 @@
-import DefaultButton from "@/components/ui/defaultButton";
-
-interface BoardPaginationProps {
-  currentPage: number;
-  totalItems: number;
-  itemsPerPage: number;
-  onPageChange: (page: number) => void;
-}
-
-export default function BoardPagination({
-  currentPage,
-  totalItems,
-  itemsPerPage,
-  onPageChange, //페이지 변경 콜백 함수
-}: BoardPaginationProps) {
-  const totalPages = Math.ceil(totalItems / itemsPerPage);
-
-  const handlePageChange = (page: number) => {
-    onPageChange(page);
-  };
-
-  const getPageNumbers = () => {
-    const pageNumbers = [];
-    const maxVisiblePages = 5; // 최대 표시할 페이지 수
-    const halfVisible = Math.floor(maxVisiblePages / 2);
-    let startPage = Math.max(1, currentPage - halfVisible);
-    let endPage = Math.min(totalPages, currentPage + halfVisible);
-    if (totalPages <= maxVisiblePages) {
-      startPage = 1;
-      endPage = totalPages;
-    } else {
-      if (currentPage <= halfVisible) {
-        endPage = maxVisiblePages;
-      } else if (currentPage + halfVisible >= totalPages) {
-        startPage = totalPages - maxVisiblePages + 1;
-      }
-    }
-    for (let i = startPage; i <= endPage; i++) {
-      pageNumbers.push(i);
-    }
-    if (startPage > 1) {
-      pageNumbers.unshift("...");
-      pageNumbers.unshift(1);
-    }
-    if (endPage < totalPages) {
-      pageNumbers.push("...");
-      pageNumbers.push(totalPages);
-    }
-
-    return pageNumbers;
-  };
-
-  const pageNumbers = getPageNumbers();
-
-  return (
-    <div className="flex items-center justify-center gap-2 mt-8">
-      <DefaultButton
-        variant="outline"
-        size="sm"
-        onClick={() => handlePageChange(currentPage - 1)}
-        disabled={currentPage === 1}
-        className="border-gray-300 text-gray-600 hover:border-cert-red disabled:opacity-50 disabled:cursor-not-allowed"
-      >
-        <svg className="w-4 h-4" viewBox="0 0 24 24" fill="none">
-          <path d="M15 18L9 12L15 6" stroke="currentColor" strokeWidth="2" />
-        </svg>{" "}
-      </DefaultButton>
-      {pageNumbers.map((page, index) => {
-        if (page === "...") {
-          return (
-            <span key={`ellipsis-${index}`} className="px-3 py-2 text-gray-400">
-              ...
-            </span>
-          );
-        }
-
-        return (
-          <DefaultButton
-            key={page}
-            variant={page === currentPage ? "default" : "outline"}
-            size="sm"
-            onClick={() => handlePageChange(page as number)}
-            className={
-              page === currentPage
-                ? "bg-cert-red text-white hover:bg-cert-red/80 min-w-[40px]"
-                : "border-gray-300 text-gray-600 hover:border-cert-red min-w-[40px]"
-            }
-          >
-            {page}
-          </DefaultButton>
-        );
-      })}
-      <DefaultButton
-        variant="outline"
-        size="sm"
-        onClick={() => handlePageChange(currentPage + 1)}
-        disabled={currentPage === totalPages}
-        className="border-gray-300 text-gray-600 hover:border-cert-red disabled:opacity-50 disabled:cursor-not-allowed"
-      >
-        <svg className="w-4 h-4" viewBox="0 0 24 24" fill="none">
-          <path d="M9 18L15 12L9 6" stroke="currentColor" strokeWidth="2" />
-        </svg>{" "}
-      </DefaultButton>
-    </div>
-  );
-}
